add basic routing with new redirect url

+++ src/App.tsx
@@ -2,14 +2,15 @@ import React from 'react';
 import Spotify, { hostname, Storage, wrapObj } from './Spotify';
 import './App.css';
 
+import { BrowserRouter as Router, Switch, Route, Redirect } from 'react-router-dom';
+
 import PlaylistExplorer from './Components/PlaylistExplorer';
 
 const spotify = wrapObj(new Spotify('104889eeeb724a9ca5efa673f527f38f'));
 export const SpotifyContext = React.createContext(spotify);
 
 const App: React.FC = () => {
-	// move to redirect page
-	if (window.location.hash) {
+	if (window.location.hash && window.location.href.includes('redirect')) {
 		const params = new URLSearchParams(window.location.hash);
 		if (params.get('state') === Storage.state) {
 			Storage.removeState();
@@ -26,8 +27,32 @@ const App: React.FC = () => {
 	return (
 		<div className="App">
 			<SpotifyContext.Provider value={spotify}>
-				{!Storage.accessToken && <button onClick={spotify.login}>login</button>}
-				<PlaylistExplorer />
+				<Router>
+					<Switch>
+						<Route path="/home">
+							{!Storage.accessToken && (
+								<button onClick={() => spotify.login(window.location.origin + '/redirect')}>
+									login
+								</button>
+							)}
+						</Route>
+						<Route path="/login"></Route>
+						<Route path="/user">
+							<div>test</div>
+						</Route>
+						<Route path="/listening"></Route>
+						<Route path="/playlists">
+							<PlaylistExplorer />
+						</Route>
+						<Route path="/redirect">
+							{/* add redirect page and redirect programmatically after timeout */}
+							{/* <Redirect push to={'home'} /> */}
+						</Route>
+						<Route>
+							<Redirect push to={'home'} />
+						</Route>
+					</Switch>
+				</Router>
 			</SpotifyContext.Provider>
 		</div>
 	);
